[proliferate] Move small chat hooks into folders
