+++ desktop/src/components/workspace/chat/ChatView.tsx
@@ -18,9 +18,9 @@ import {
 import { useChatAvailabilityState } from "@/hooks/chat/derived/use-chat-availability-state";
 import { useChatDockInset } from "@/hooks/chat/ui/use-chat-dock-inset";
 import { useChatPromptAttachments } from "@/hooks/chat/ui/use-chat-prompt-attachments";
-import { useCloudWorkspacePolling } from "@/hooks/chat/use-cloud-workspace-polling";
+import { useCloudWorkspacePolling } from "@/hooks/chat/lifecycle/use-cloud-workspace-polling";
 import { useComposerDockSlots } from "@/hooks/chat/ui/use-composer-dock-slots";
-import { useQueuedPromptEditStatus } from "@/hooks/chat/use-queued-prompt-edit";
+import { useQueuedPromptEditStatus } from "@/hooks/chat/workflows/use-queued-prompt-edit";
 import { useDebugRenderCount } from "@/hooks/ui/use-debug-render-count";
 import { useDebugValueChange } from "@/hooks/ui/use-debug-value-change";
 import { useSessionErrorAcknowledgement } from "@/hooks/sessions/lifecycle/use-session-error-acknowledgement";
