+++ desktop/src/components/workspace/chat/input/ApprovalCard.tsx
@@ -1,7 +1,7 @@
 import { Button } from "@/components/ui/Button";
 import { ComposerAttachedPanel } from "./ComposerAttachedPanel";
 import { useActivePendingApproval } from "@/hooks/chat/derived/use-active-chat-session-selectors";
-import { useChatPermissionActions } from "@/hooks/chat/use-chat-permission-actions";
+import { useChatPermissionActions } from "@/hooks/chat/workflows/use-chat-permission-actions";
 import type { PermissionOptionAction } from "@/lib/domain/chat/composer/chat-input-helpers";
 
 // Codex's approval header is just the title as medium-weight text — no
