+++ desktop/src/components/workspace/chat/input/PendingPromptList.tsx
@@ -3,7 +3,7 @@ import { Button } from "@/components/ui/Button";
 import { Pencil, X } from "@/components/ui/icons";
 import { useActiveSessionId } from "@/hooks/chat/derived/use-active-chat-session-selectors";
 import { usePromptOutboxActions } from "@/hooks/chat/workflows/use-prompt-outbox-actions";
-import { useQueuedPromptEditReader } from "@/hooks/chat/use-queued-prompt-edit";
+import { useQueuedPromptEditReader } from "@/hooks/chat/workflows/use-queued-prompt-edit";
 import { useDeletePendingPrompt } from "@/hooks/sessions/workflows/use-delete-pending-prompt";
 import {
   derivePendingPromptQueueRow,
