+++ desktop/src/components/workspace/chat/input/delegated-work/AgentsPopoverSubagentSection.tsx
@@ -1,6 +1,6 @@
 import { Button } from "@/components/ui/Button";
 import { ExternalLink } from "@/components/ui/icons";
-import type { DelegatedWorkComposerViewModel } from "@/hooks/chat/use-delegated-work-composer";
+import type { DelegatedWorkComposerViewModel } from "@/hooks/chat/facade/use-delegated-work-composer";
 import { PopoverSection } from "./PopoverSection";
 
 export function AgentsPopoverSubagentSection({
