+++ desktop/src/hooks/chat/lifecycle/use-cloud-workspace-polling.ts
