+++ desktop/src/hooks/chat/ui/use-chat-draft-state.ts
