+++ desktop/src/hooks/chat/ui/use-composer-dock-slots.tsx
@@ -13,7 +13,7 @@ import {
   useActivePendingPrompts,
 } from "@/hooks/chat/derived/use-active-chat-session-selectors";
 import { useDebugValueChange } from "@/hooks/ui/use-debug-value-change";
-import { useDelegatedWorkComposer } from "@/hooks/chat/use-delegated-work-composer";
+import { useDelegatedWorkComposer } from "@/hooks/chat/facade/use-delegated-work-composer";
 import { useActiveTodoTracker } from "@/hooks/chat/derived/use-active-todo-tracker";
 import { useSelectedCloudRuntimeState } from "@/hooks/workspaces/use-selected-cloud-runtime-state";
 import { useWorkspaceStatusPanelState } from "@/hooks/workspaces/derived/use-workspace-status-panel-state";
