+++ desktop/src/hooks/chat/workflows/use-chat-launch-intent-actions.ts
