+++ desktop/src/hooks/chat/workflows/use-chat-mcp-elicitation-actions.ts
