+++ desktop/src/hooks/chat/workflows/use-queued-prompt-edit.ts
